requirements.txt
langchain
google.generativeai
huggingface_hub
python-dotenv
streamlit